fix: 선택된 날짜와 부스 리스트 동기화 안되는 문제 해결
